backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0

# AI & Agents
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
langgraph>=0.3,<0.4
langsmith>=0.3,<0.4

# Database & Vector Store
supabase==2.3.6
sqlalchemy[asyncio]==2.0.47
alembic==1.15.2
asyncpg==0.30.0           # Async PostgreSQL driver
pgvector==0.3.6           # pgvector Python bindings
psycopg2-binary==2.9.11   # Sync driver for Alembic migrations

# Authentication & external
python-jose==3.3.0        # For Clerk JWT validation
stripe==8.5.0
httpx==0.25.2             # For async HTTP requests (Firecrawl, Tavily etc.)

# Utilities
python-dotenv==1.0.1
reportlab>=4.0,<5.0         # PDF export for reports
